#78: refactor tests helpers into separate files
