+++ src/api/tests/models/test_run_parameters.py
@@ -1,8 +1,7 @@
 from django.test import TestCase
 
-from api.models.models import LAWMRunParameters, LAWMGeneralParameters, LAWMRegionalParameters, LAWMRegion
-from api.std_lib.lawm.regions import DEFAULT_REGIONS
-from api.tests.api_test_mixin import MicroSimuTestMixin
+from api.models.models import LAWMRunParameters, LAWMGeneralParameters, LAWMRegionalParameters
+from api.tests.helpers.api_test_mixin import MicroSimuTestMixin
 
 
 class LAWMRunParametersTest(TestCase, MicroSimuTestMixin):
