+++ src/api/tests/serializers/test_many_regional_parameters_serializer.py
@@ -1,7 +1,7 @@
 from django import test
 
 from api.serializers.parameters_serializers import RegionalParametersSerializer, ManyRegionalParametersSerializer
-from api.tests.api_test_mixin import MicroSimuTestMixin
+from api.tests.helpers.api_test_mixin import MicroSimuTestMixin
 
 
 class RegionalParametersListSerializerTest(test.TestCase, MicroSimuTestMixin):
