+++ src/api/tests/serializers/test_region_result_serializer.py
@@ -4,7 +4,7 @@ from django.urls import path, include, reverse
 
 from api.serializers.results_serializers import RegionResultSerializer
 from api.std_lib.lawm.variables import Population
-from api.tests.api_test_mixin import MicroSimuTestMixin
+from api.tests.helpers.api_test_mixin import MicroSimuTestMixin
 
 test_patterns = [
     path('test_path/<int:pk>/'             , lambda request: None, name='simulation-detail'),
