+++ src/api/tests/serializers/test_regional_parameters_serializer.py
@@ -2,7 +2,7 @@ from django import test
 
 from api.models.models import LAWMRegionalParameters
 from api.serializers.parameters_serializers import RegionalParametersSerializer
-from api.tests.api_test_mixin import MicroSimuTestMixin
+from api.tests.helpers.api_test_mixin import MicroSimuTestMixin
 
 
 class RegionalParametersSerializerTest(test.TestCase, MicroSimuTestMixin):
