+++ src/api/tests/serializers/test_result_serializer.py
@@ -1,7 +1,7 @@
 from django import test
 
 from api.serializers.results_serializers import ResultSerializer
-from api.tests.api_test_mixin import MicroSimuTestMixin
+from api.tests.helpers.api_test_mixin import MicroSimuTestMixin
 
 
 class ResultSerializerTest(test.TestCase, MicroSimuTestMixin):
